+++ frontend/components/admin/KeywordControl.tsx
@@ -1,0 +1,9 @@
+const KeywordControl = () => {
+	return (
+		<>
+			<h4>키워드 관리</h4>
+		</>
+	);
+};
+
+export default KeywordControl;
