+++ frontend/components/admin/UserControl.tsx
@@ -1,0 +1,9 @@
+const UserControl = () => {
+	return (
+		<>
+			<h4>유저 관리</h4>
+		</>
+	);
+};
+
+export default UserControl;
