+++ frontend/components/admin/index.tsx
@@ -1,0 +1,5 @@
+import AdminMainLayout from './AdminMainLayout';
+import KeywordControl from './KeywordControl';
+import UserControl from './UserControl';
+
+export { AdminMainLayout, KeywordControl, UserControl };
